+++ changes.md
@@ -53,6 +53,8 @@
 <details>
 <summary>Added Resource(s)</summary>
 
+- added resource `/{projectKey}/customers/search`
+- added resource `/{projectKey}/customers/search/indexing-status`
 - added resource `/{projectKey}/in-store/key={storeKey}/quote-requests`
 - added resource `/{projectKey}/in-store/key={storeKey}/staged-quotes`
 - added resource `/{projectKey}/in-store/key={storeKey}/quotes`
@@ -86,6 +88,9 @@
 <details>
 <summary>Added Method(s)</summary>
 
+- added method `$apiRoot->withProjectKey()->customers()->search()->post()`
+- added method `$apiRoot->withProjectKey()->customers()->search()->head()`
+- added method `$apiRoot->withProjectKey()->customers()->searchIndexingStatus()->get()`
 - added method `$apiRoot->withProjectKey()->inStoreKeyWithStoreKeyValue()->quoteRequests()->get()`
 - added method `$apiRoot->withProjectKey()->inStoreKeyWithStoreKeyValue()->quoteRequests()->head()`
 - added method `$apiRoot->withProjectKey()->inStoreKeyWithStoreKeyValue()->quoteRequests()->post()`
@@ -137,6 +142,14 @@
 
 - added type `Attribution`
 - added type `AttributionSource`
+- added type `CustomerIndexingProgress`
+- added type `CustomerIndexingStatus`
+- added type `CustomerPagedSearchResponse`
+- added type `CustomerSearchIndexingStatusResponse`
+- added type `CustomerSearchRequest`
+- added type `CustomerSearchResult`
+- added type `CustomerSearchProjectNotIndexedError`
+- added type `GraphQLCustomerSearchProjectNotIndexedError`
 - added type `ProductTailoringImageAddedMessage`
 - added type `ProductTailoringImagesSetMessage`
 - added type `ProductVariantTailoringAddedMessage`
@@ -168,6 +181,8 @@
 - added type `ProductTailoringSetAssetTagsAction`
 - added type `ProductTailoringSetExternalImagesAction`
 - added type `ProductTailoringSetImageLabelAction`
+- added type `CustomerSearchStatus`
+- added type `ProjectChangeCustomerSearchStatusAction`
 </details>